# backend/auth-facade/openapi.yaml
openapi: 3.1.0
info:
  title: Auth Facade API
  description: |
    Backend for Frontend (BFF) authentication service for the EMS platform.

    ## Overview
    The Auth Facade provides provider-agnostic authentication, abstracting identity providers
    (Keycloak, Auth0, Okta, Azure AD) behind a unified API. This allows switching identity
    providers via YAML configuration without code changes.

    ## Architecture
    - **Strategy Pattern**: Identity providers implement the `IdentityProvider` interface
    - **Provider Selection**: Set `auth.facade.provider` in configuration (keycloak, auth0, okta, azure-ad)
    - **Multi-tenancy**: Tenant ID passed via `X-Tenant-ID` header, maps to IdP realm/tenant

    ## Authentication Flows
    1. **Direct Login**: Email/password via Resource Owner Password Credentials
    2. **Social Login**: Google One Tap, Microsoft MSAL token exchange
    3. **Provider Redirect**: OAuth2/OIDC redirect via `kc_idp_hint` (dynamic provider selection)
    4. **MFA**: TOTP-based two-factor authentication

    ## Token Management
    - Access tokens: Short-lived JWT (5-15 minutes)
    - Refresh tokens: Longer-lived, with rotation on use
    - Token blacklisting via Valkey for logout/revocation
  version: 1.0.0
  contact:
    name: EMS Team
    email: support@ems.com
  license:
    name: Proprietary
    url: https://ems.com/license
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://auth.ems.com
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/me:
    get:
      tags: [User Profile]
      summary: Get current user profile
      description: |
        Get profile information for the authenticated user.

        User information is extracted from the JWT claims using the
        provider-agnostic claim mapping configured in `auth.facade.user-claim-mappings`.
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                email: "user@example.com"
                firstName: "John"
                lastName: "Doe"
                tenantId: "tenant-acme"
                roles:
                  - "ROLE_USER"
                  - "ROLE_MANAGER"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "unauthorized"
                message: "Not authenticated"
                timestamp: "2024-01-15T10:30:00Z"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from login endpoints.

        **Header format:** `Authorization: Bearer <token>`

        **Token structure:**
        - `sub`: User ID
        - `email`: User email
        - `tenant_id`: Tenant identifier
        - `realm_access.roles`: User roles (Keycloak)
        - `exp`: Expiration timestamp
...
  schemas:
...
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier (from IdP `sub` claim)
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        firstName:
          type: string
          description: User's first name (given_name claim)
          example: "John"
        lastName:
          type: string
          description: User's last name (family_name claim)
          example: "Doe"
        tenantId:
          type: string
          description: Tenant identifier the user belongs to
          example: "tenant-acme"
        roles:
          type: array
          items:
            type: string
          description: |
            User's roles/authorities. Extracted from IdP using
            configured role claim paths.
          example:
            - "ROLE_USER"
            - "ROLE_MANAGER"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          example: "invalid_credentials"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid email or password"
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-01-15T10:30:00Z"
        details:
          type: object
          additionalProperties: true
          description: Additional error details (optional)